feat: add React Router with URL-based navigation
- Install @tanstack/router and react-router
- Create src/routes/ with lazy-loaded page components:
  - PresetsPage (/presets)
  - PerformancePage (/performance)
  - SynthBackupsPage (/synth-backups)
  - SetlistsPage (/setlists)
  - TagManagerPage (/tags)
  - DuplicateFinderPage (/duplicates)
- Migrate AppMode state to URL params
- Remove old PresetManager.tsx (replaced by routes)
- AppSidebar accepts onNavigate callback instead of setAppMode
- All routes lazy-loaded with Suspense for code splitting
- Performance mode preserves selected preset via ?presetId query param
- Add getPresetById() helper function
