+++ package.json
@@ -30,6 +30,7 @@
 		"@tanstack/react-query": "^5.96.2",
 		"@tanstack/react-table": "^8.21.3",
 		"@tanstack/react-virtual": "^3.13.23",
+		"@tanstack/router": "^0.0.1-beta.53",
 		"@tauri-apps/api": "^2.10.1",
 		"@tauri-apps/plugin-dialog": "~2.7.0",
 		"@tauri-apps/plugin-fs": "~2.5.0",
@@ -41,6 +42,7 @@
 		"react": "^18.3.1",
 		"react-dom": "^18.3.1",
 		"react-icons": "^5.6.0",
+		"react-router": "^7.14.0",
 		"react-virtualized-auto-sizer": "^1.0.26",
 		"react-window": "^1.8.11",
 		"react-window-infinite-loader": "^1.0.10",
